printing vec and size check

+++ src/map_builder.rs
@@ -1,5 +1,5 @@
-pub fn map_builder() -> Vec<u8> {
-  let mut v: Vec<u8> = Vec::new();
+pub fn map_builder() -> Vec<u32> {
+  let mut v: Vec<u32> = Vec::new();
   v.reserve(81);
   v
 }
